DualNumbers.jl

DualNumbers.Dual — Type

Dual numbers with a scalar dual part

DualNumbers.derivative — Function

derivative(f, x)
derivative(f, xs, i)
derivative(f, x, c)
derivative(f, xs, i, c)

Evaluate the partial derivative of the scalar-argument function f(x) at x, or of the array-argument function f(xs) at xs[i].

When the argument x (or xs[i]) is complex, then an additional argument c chooses whether the partial derivative is with respect to the real (c==1) or imaginary (c==2) part of the argument.